+++ vms/platformvm/config/config.go
@@ -39,7 +39,7 @@ type Config struct {
 	ChainDBCacheSize              int           `json:"chain-db-cache-size"`
 	BlockIDCacheSize              int           `json:"block-id-cache-size"`
 	FxOwnerCacheSize              int           `json:"fx-owner-cache-size"`
-	SubnetToL1ConversionCacheSize int           `json:"subnet-conversion-cache-size"`
+	SubnetToL1ConversionCacheSize int           `json:"subnet-to-l1-conversion-cache-size"`
 	L1WeightsCacheSize            int           `json:"l1-weights-cache-size"`
 	L1InactiveValidatorsCacheSize int           `json:"l1-inactive-validators-cache-size"`
 	L1SubnetIDNodeIDCacheSize     int           `json:"l1-subnet-id-node-id-cache-size"`