+++ RELEASES.md
@@ -5,6 +5,7 @@
 ### APIs
 
 - Added `platform.getL1Validator`
+- Added `platform.getProposedHeight`
 
 ### Configs